python/requirements.txt
py-openaq
fastapi
uvicorn
numpy
pandas
scikit-learn